+++ packages/components-native/src/__mocks__/__mocks.ts
@@ -1,11 +1,29 @@
 import { MissingTranslationError } from "react-intl";
-// eslint-disable-next-line import/no-internal-modules
 import mockRNLocalize from "react-native-localize/mock";
 
 jest.mock("react-native-localize", () => ({
   ...mockRNLocalize,
   getTimeZone: () => "UTC",
 }));
+
+jest.mock("react-native-reanimated", () => {
+  const reanimated = require("react-native-reanimated/mock");
+  const timing = () => ({ start: () => undefined });
+  return {
+    ...reanimated,
+    default: {
+      ...reanimated.default,
+      timing,
+      // The mock for `call` immediately calls the callback which is incorrect
+      // So we override it with a no-op
+      call: () => undefined,
+    },
+    timing,
+    FadeIn: {
+      duration: () => undefined,
+    },
+  };
+});
 
 export const MOCK_INTL_CONFIG = {
   locale: "en-US",